Move split database options to a database model
