+++ infer.py
@@ -2,7 +2,7 @@
 
 import tensorflow as tf
 
-from microlensing_label_per_example_database import MicrolensingLabelPerExampleDatabase
+from photometric_database.microlensing_label_per_example_database import MicrolensingLabelPerExampleDatabase
 from models import SimpleLightcurveCnn
 
 print('Preprocessing data...')
