+++ photometric_database/microlensing_label_per_example_database.py
@@ -6,7 +6,7 @@ import numpy as np
 import pandas as pd
 import tensorflow as tf
 
-from lightcurve_database import LightcurveDatabase
+from photometric_database.lightcurve_database import LightcurveDatabase
 
 
 class MicrolensingLabelPerExampleDatabase(LightcurveDatabase):
